Move ICDS database to ICDS directory
