+++ corehq/sql_db/tests/test_routers.py
@@ -59,7 +59,7 @@ class AllowMigrateTest(SimpleTestCase):
         self.assertIs(False, allow_migrate('default', SYNCLOGS_APP))
         self.assertIs(True, allow_migrate('synclogs', SYNCLOGS_APP))
 
-    @patch('corehq.sql_db.routers.get_icds_ucr_db_alias')
+    @patch('custom.icds_reports.router.get_icds_ucr_db_alias')
     def test_icds_db(self, mock):
         mock.return_value = None
         self.assertIs(False, allow_migrate('default', ICDS_REPORTS_APP))
