+++ custom/icds_reports/management/commands/imbalanced_indicator.py
@@ -7,7 +7,7 @@ import six
 from django.core.management.base import BaseCommand
 from django.db import connections
 
-from corehq.sql_db.connections import get_icds_ucr_db_alias
+from custom.icds_reports.router import get_icds_ucr_db_alias
 
 
 class Command(BaseCommand):
