+++ custom/icds_reports/management/commands/update_child_health_monthly.py
@@ -9,7 +9,7 @@ from django.core.management.base import BaseCommand
 
 from django.db import connections, transaction
 
-from corehq.sql_db.connections import get_icds_ucr_db_alias
+from custom.icds_reports.router import get_icds_ucr_db_alias
 from io import open
