+++ custom/icds_reports/tests/__init__.py
@@ -19,7 +19,8 @@ from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.locations.models import SQLLocation, LocationType
 from corehq.apps.userreports.models import StaticDataSourceConfiguration
 from corehq.apps.userreports.util import get_indicator_adapter, get_table_name
-from corehq.sql_db.connections import connection_manager, ICDS_UCR_ENGINE_ID
+from corehq.sql_db.connections import connection_manager
+from custom.icds_reports.const import ICDS_UCR_ENGINE_ID
 from custom.icds_reports.tasks import (
     move_ucr_data_into_aggregation_tables,
     build_incentive_report,
